fix type for Sound and Note components

+++ compiler/library/blg-mb-1-a7/1.0/source/SpeakerNote.hpp
@@ -30,7 +30,7 @@ public:
 
     // TODO: duration in ms - DEFAULT PARAM
     // TODO:В ПЛАТФОРМЕ ЕГО НЕ ДОЛЖНО БЫТЬ
-    void play(Note *note, const uint16_t duration) {
+    void play(Note *note, const uint32_t duration) {
 
         // TODO: Если не работает, то вернуть тип на тип Sound* и не парить мозги
         mrx::hal::speaker::startSound(reinterpret_cast<Sound*>(note), duration);
